Remove uint8 conversion warning

+++ src/view.cpp
@@ -168,7 +168,7 @@ void MainView::render_text(std::string text, Color cl, int fontSize, std::string
     }
 
     //
-    SDL_Color color = {cl.r, cl.g, cl.b, cl.a};
+    SDL_Color color = {(Uint8)cl.r, (Uint8)cl.g, (Uint8)cl.b, (Uint8)cl.a};
     // SDL_Surface* surface = TTF_RenderText_Solid(dynamicFont, text.c_str(), color);     // No anti-aliasing (blocky text).
     // SDL_Surface* surface = TTF_RenderText_Shaded(dynamicFont, text.c_str(), color);     // Anti-aliasing with background color.
     SDL_Surface* surface = TTF_RenderText_Blended(dynamicFont, text.c_str(), color);     // Best anti-aliasing, recommended for smooth text.
